fix(core): use COPYFILE_EXCL when installing native binary
This will prevent a race condition when multiple nx packages of the same
version are installed in parallel. For example, if a workspace has nx
v19.0.0 installed in package.json, but is using ^19.0.0 for plugin
versions, the plugins will install separate copies of nx in parallel.
This could cause the native temp file to get clobbered and corrupted.

This change adds the COPYFILE_EXCL mode flag to the copy command, which
will open the destination file with the O_EXCL flag. This ensures the
destination file is not open by any other process at the same time.