fix(core): use a retry loop when writing temp native bindings file
This avoids another race condition where another process has started
writing the file, but has not finished before we try to read it.

If the native file is required again later, node will first use the require cache.
If it's a new node process requiring the file, this will load the first iteration
that was successfully written by any other process.